Task 6

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -1,4 +1,3 @@
-#include <stdlib.h>
 #include <stdio.h>
 /**
 *main - entry point, print the alphabets
@@ -8,7 +7,7 @@
 */
 int main(void)
 {
-	int i = 0;
+	long int i = 0;
 	while (i < 10)
 	{
 		putchar(i);
